+++ src/utils/prisma.ts
@@ -1,10 +1,39 @@
-import { ChatSettingName, LanguageCode, Prisma, PrismaClient, User } from "@prisma/client";
+import {
+  AddingBotsRule,
+  ChatSettingName,
+  ChatType,
+  LanguageCode,
+  Prisma,
+  PrismaClient,
+  ProfanityFilterRule,
+  User,
+} from "@prisma/client";
 import { formatInTimeZone } from "date-fns-tz";
 import i18next, { t } from "i18next";
 import { Chat as TelegramChat, User as TelegramUser } from "telegraf/typings/core/types/typegram";
 import { PrismaChat } from "types/prismaChat";
 import { DATE_FORMAT, DATE_LOCALES } from "utils/consts";
 import { bot, getChatDisplayTitle, getUserDisplayName, getUserHtmlLink } from "utils/telegraf";
+
+/**
+ * Resolves chat type based on Telegram chat type
+ * @param chatType Telegram chat type
+ * @returns Chat type which is supported by the bot
+ */
+const resolveChatType = (chatType: TelegramChat["type"]): ChatType => {
+  switch (chatType) {
+    case "channel":
+      return ChatType.CHANNEL;
+    case "group":
+      return ChatType.GROUP;
+    case "private":
+      return ChatType.PRIVATE;
+    case "supergroup":
+      return ChatType.SUPERGROUP;
+    default:
+      return ChatType.UNKNOWN;
+  }
+};
 
 /**
  * Resolves language based on Telegram language code
@@ -14,10 +43,10 @@ import { bot, getChatDisplayTitle, getUserDisplayName, getUserHtmlLink } from "u
 const resolveLanguage = (languageCode: string | undefined): LanguageCode => {
   switch (languageCode) {
     case "ru":
-      return "ru";
+      return LanguageCode.RU;
     case "en":
     default:
-      return "en";
+      return LanguageCode.EN;
   }
 };
 
@@ -112,7 +141,7 @@ export const upsertPrismaChat = async (chat: TelegramChat, editor: TelegramUser)
       .map((u) => upsertPrismaUser(u, editor)),
     prisma.chat.upsert({
       create: {
-        addingBots: ["group", "supergroup"].includes(chat.type) ? "restricted" : undefined,
+        addingBots: chat.type === "group" || chat.type === "supergroup" ? AddingBotsRule.RESTRICT : undefined,
         admins: { connect: admins.map((a) => ({ id: a.user.id })) },
         authorId: editor.id,
         displayTitle,
@@ -122,10 +151,10 @@ export const upsertPrismaChat = async (chat: TelegramChat, editor: TelegramUser)
         language: resolveLanguage(editor.language_code),
         lastName,
         membersCount,
-        profanityFilter: ["group", "supergroup"].includes(chat.type) ? "enabled" : undefined,
+        profanityFilter: chat.type === "group" || chat.type === "supergroup" ? ProfanityFilterRule.FILTER : undefined,
         timeZone: resolveTimeZone(editor.language_code),
         title,
-        type: chat.type,
+        type: resolveChatType(chat.type),
         username,
       },
       include: { admins: true, chatSettingsHistory: { include: { editor: true } } },
@@ -137,7 +166,7 @@ export const upsertPrismaChat = async (chat: TelegramChat, editor: TelegramUser)
         lastName,
         membersCount,
         title,
-        type: chat.type,
+        type: resolveChatType(chat.type),
         username,
       },
       where: { id: chat.id },
@@ -149,10 +178,10 @@ export const upsertPrismaChat = async (chat: TelegramChat, editor: TelegramUser)
     throw new Error("Something went wrong during chat upsertion.");
   }
 
-  // Patch bot private chat title
-  const patchedDisplayTitle =
-    bot.botInfo && chat.id === editor.id ? getUserDisplayName(bot.botInfo, "full") : displayTitle;
-  return { ...prismaChat, displayTitle: patchedDisplayTitle };
+  return bot.botInfo && chat.id === editor.id
+    ? // Patch display title of the chat with the bot
+      { ...prismaChat, displayTitle: getUserDisplayName(bot.botInfo, "full"), username: bot.botInfo.username }
+    : prismaChat;
 };
 
 /**
@@ -195,11 +224,11 @@ export const upsertPrismaSenderChat = async (chat: TelegramChat, editor: Telegra
         id: chat.id,
         lastName,
         title,
-        type: chat.type,
+        type: resolveChatType(chat.type),
         username,
       },
       select: { id: true },
-      update: { editorId: editor.id, firstName, lastName, title, type: chat.type, username },
+      update: { editorId: editor.id, firstName, lastName, title, type: resolveChatType(chat.type), username },
       where: { id: chat.id },
     }),
   ]);