+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.10.2 (2023-11-12)
+
+- Fix: incorrect display of the name of the chat with the bot
+- Other: internal improvements
+
 ## 1.10.1 (2023-11-12)
 
 - Fix: the chat name is not displayed correctly if it contains "&", "<" or ">" symbols